AAS_send/requirements.txt
paho-mqtt==1.6.1
chevron==0.14.0
tomli==2.0.1